feat: PolynomialRingElement trait, polynomial_ring macro
