+++ ring-math/src/polynomial_ring.rs
@@ -1,0 +1,320 @@
+use std::fmt::Debug;
+use std::fmt::Display;
+use std::hash::Hash;
+use std::ops::Add;
+use std::ops::AddAssign;
+use std::ops::Div;
+use std::ops::Mul;
+use std::ops::MulAssign;
+use std::ops::Neg;
+use std::ops::Sub;
+use std::ops::SubAssign;
+use std::str::FromStr;
+
+use scalarff::FieldElement;
+
+use super::polynomial::Polynomial;
+
+/// A trait representing a polynomial ring
+/// defined as `T[X]/<Self::modulus()>`
+/// where T is a FieldElement trait
+/// and modulus is a function implemented by the
+/// struct
+pub trait PolynomialRingElement<T: FieldElement>:
+    Add<Output = Self>
+    + AddAssign
+    + Div<Output = Self>
+    + Mul<Output = Self>
+    + MulAssign
+    + Neg<Output = Self>
+    + Sub<Output = Self>
+    + SubAssign
+    + FromStr
+    + PartialEq
+    + Clone
+    + Hash
+    + Debug
+    + From<u64>
+    + Display
+{
+    /// Modulus used in remainder division to form
+    /// the polynomial ring.
+    ///
+    /// Operations are done modulo this polynomial,
+    /// in a way similar to scalar fields.
+    ///
+    /// See the division implementation for more info.
+    fn modulus() -> Polynomial<T>;
+
+    /// Return the Polynomial representation of the current value
+    /// Used to automatically implement norms and other functions.
+    fn polynomial(&self) -> &Polynomial<T>;
+
+    /// Calculate the l1 norm for this polynomial. That is
+    /// the summation of all coefficients
+    fn norm_l1(&self) -> u64 {
+        let digits = self
+            .polynomial()
+            .coefficients
+            .iter()
+            .fold(T::zero(), |acc, x| acc + x.clone())
+            .to_biguint()
+            .to_u64_digits();
+        if digits.len() > 1 {
+            panic!("Norm l1 is not a single u64 digit");
+        } else if digits.len() == 1 {
+            digits[0]
+        } else {
+            0
+        }
+    }
+
+    /// Calculate the l2 norm for this polynomial. That is
+    /// the square root of the summation of each coefficient squared
+    ///
+    /// Specifically, we're calculating the square root in the integer
+    /// field, not the prime field
+    fn norm_l2(&self) -> u64 {
+        let v = self
+            .polynomial()
+            .coefficients
+            .iter()
+            .fold(T::zero(), |acc, x| acc + (x.clone() * x.clone()));
+        let digits = v.to_biguint().sqrt().to_u64_digits();
+        if digits.len() > 1 {
+            panic!("Norm l2 is not a single u64 digit");
+        } else if digits.len() == 1 {
+            digits[0]
+        } else {
+            0
+        }
+    }
+
+    /// Calculate the l-infinity norm for this polynomial. That is
+    /// the largest coefficient
+    fn norm_max(&self) -> u64 {
+        let mut max = T::zero().to_biguint();
+        for i in &self.polynomial().coefficients {
+            if i.to_biguint() > max {
+                max = i.to_biguint();
+            }
+        }
+        let digits = max.to_u64_digits();
+        if digits.len() > 1 {
+            panic!("Norm max is not a single u64 digit");
+        } else if digits.len() == 1 {
+            digits[0]
+        } else {
+            0
+        }
+    }
+}
+
+#[macro_export]
+macro_rules! polynomial_ring {
+    ( $name: ident, $modulus: expr, $name_str: expr ) => {
+        type T = FoiFieldElement;
+
+        #[derive(Clone, Debug, PartialEq, Eq, std::hash::Hash)]
+        pub struct $name(Polynomial<T>);
+
+        impl PolynomialRingElement<T> for $name {
+            fn modulus() -> Polynomial<T> {
+                $modulus
+            }
+            fn polynomial(&self) -> &Polynomial<T> {
+                &self.0
+            }
+        }
+
+        impl From<Polynomial<T>> for $name {
+            fn from(p: Polynomial<T>) -> Self {
+                $name(p.div(&Self::modulus()).1)
+            }
+        }
+
+        impl FieldElement for $name {
+            fn zero() -> Self {
+                $name(Polynomial {
+                    coefficients: vec![T::zero()],
+                })
+            }
+
+            fn one() -> Self {
+                $name(Polynomial::identity())
+            }
+
+            fn byte_len() -> usize {
+                Self::modulus().degree() * T::byte_len()
+            }
+
+            fn serialize(&self) -> String {
+                self.0
+                    .coefficients
+                    .iter()
+                    .map(|v| v.serialize())
+                    .collect::<Vec<_>>()
+                    .join(",")
+            }
+
+            fn deserialize(str: &str) -> Self {
+                $name(Polynomial {
+                    coefficients: str
+                        .split(',')
+                        .map(|v| T::deserialize(v))
+                        .collect::<Vec<_>>(),
+                })
+            }
+
+            fn prime() -> BigUint {
+                panic!("cannot retrieve a scalar prime for a polynomial field");
+            }
+
+            fn name_str() -> &'static str {
+                "polynomial_ring"
+            }
+
+            /// Return a constant polynomial with the provided
+            /// value
+            fn from_usize(value: usize) -> Self {
+                $name(Polynomial {
+                    coefficients: vec![T::from_usize(value)],
+                })
+            }
+
+            fn to_biguint(&self) -> BigUint {
+                panic!("cannot retrieve a scalar representation for a polynomial field element");
+            }
+
+            fn from_biguint(v: &BigUint) -> Self {
+                panic!();
+            }
+
+            fn from_bytes_le(bytes: &[u8]) -> Self {
+                $name(Polynomial {
+                    coefficients: bytes
+                        .chunks(T::byte_len())
+                        .map(|chunk| T::from_bytes_le(chunk))
+                        .collect::<Vec<_>>(),
+                })
+            }
+
+            fn to_bytes_le(&self) -> Vec<u8> {
+                self.0
+                    .coefficients
+                    .iter()
+                    .flat_map(|v| v.to_bytes_le())
+                    .collect()
+            }
+        }
+
+        impl std::fmt::Display for $name {
+            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                write!(f, "{}", self.0)
+            }
+        }
+
+        impl std::str::FromStr for $name {
+            type Err = ();
+
+            fn from_str(s: &str) -> Result<Self, Self::Err> {
+                Err(())
+            }
+        }
+
+        impl From<u64> for $name {
+            fn from(value: u64) -> Self {
+                Self::from(Polynomial {
+                    coefficients: vec![T::from(value)],
+                })
+            }
+        }
+
+        impl std::ops::Add for $name {
+            type Output = Self;
+
+            fn add(self, other: Self) -> Self {
+                Self::from(self.0 + other.0)
+            }
+        }
+
+        impl std::ops::Sub for $name {
+            type Output = Self;
+
+            fn sub(self, other: Self) -> Self {
+                Self::from(self.0 - other.0)
+            }
+        }
+
+        impl std::ops::Mul for $name {
+            type Output = Self;
+
+            fn mul(self, other: Self) -> Self {
+                Self::from(self.0 * other.0)
+            }
+        }
+
+        impl std::ops::Div for $name {
+            type Output = Self;
+
+            fn div(self, other: Self) -> Self {
+                // this implementation implies floored division, so discard the remainder
+                Self::from(self.0.div(&other.0).0)
+            }
+        }
+
+        impl std::ops::AddAssign for $name {
+            fn add_assign(&mut self, other: Self) {
+                *self = self.clone() + other;
+            }
+        }
+
+        impl std::ops::MulAssign for $name {
+            fn mul_assign(&mut self, other: Self) {
+                *self = self.clone() * other;
+            }
+        }
+
+        impl std::ops::SubAssign for $name {
+            fn sub_assign(&mut self, other: Self) {
+                *self = self.clone() - other;
+            }
+        }
+
+        impl std::ops::Neg for $name {
+            type Output = Self;
+
+            fn neg(self) -> Self {
+                $name(-self.0.clone())
+            }
+        }
+    };
+}
+
+#[cfg(test)]
+mod test {
+    use scalarff::BigUint;
+    use scalarff::FieldElement;
+    use scalarff::FoiFieldElement;
+
+    use super::Polynomial;
+    use super::PolynomialRingElement;
+
+    polynomial_ring!(
+        Poly64,
+        Polynomial::<FoiFieldElement>::new(vec![FoiFieldElement::one(), FoiFieldElement::one()]),
+        "Poly64"
+    );
+
+    #[test]
+    fn scalar_math_in_ring() {
+        for x in 100..500 {
+            for y in 200..600 {
+                let z_scalar = FoiFieldElement::from(x) * FoiFieldElement::from(y);
+                let z_poly = Poly64::from(x) * Poly64::from(y);
+                assert_eq!(z_poly.polynomial().degree(), 0);
+                assert_eq!(z_poly.polynomial().coefficients[0], z_scalar);
+            }
+        }
+    }
+}
